% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize.R
\docType{package}
\name{UniFORM-package}
\alias{UniFORM-package}
\alias{UniFORM}
\title{UniFORM Package}
\description{
An R implementation of the UniFORM pipeline for normalizing multiplex
immunofluorescence data. This package provides robust normalization methods
to harmonize intensity measurements across different samples, batches, and
experimental conditions.
}
\details{
Universal ImmunoFluorescence nORMalization
}
\keyword{internal}
